Moved logic to separate hook
