+++ components/Charts/Preview/PreviewChart.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+import { IHistory } from "../../../types/api/IHistory";
+
+interface FormattedHistory {
+  x: string;
+  y: string;
+}
+
+const PreviewChart = ({ data }: IHistory) => {};
+
+export default PreviewChart;
